Changed approach to get metrics - ScaleIO REST-API integration
- Major rewrite of the portions of code used to get metrics from the
ScaleIO infrastructure. Now the plugin queries the ScaleIO Gateway using
the provided REST-API. The bash scli wrapper is no longer needed and the
metric software can run on a server completely external to the ScaleIO
infrastructure. This server only needs to establish connection to the
ScaleIO Gateway using the HTTPS protocol

- The parameters that have to be configured in the configuration file
are changed from the original implementation because of the mentioned
redesign. Read carefully the README.md file (section: Collectd plugin
configuration)

- The grafana dashboard was left untouched